+++ Runtime/MonoEditorUtility/ShortcutListener.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using UnityEngine;
+
+#if ENABLE_INPUT_SYSTEM
 using UnityEngine.InputSystem;
+#endif
 
 namespace OcUtility
 {